Cleanup: Object - Use static tooltip description for `"Transforms to Deltas"` operator instead of dynamic property
The `"Transforms to Deltas"` has an `arg` property that allows for dynamically overriding the tooltip description. However, in all places where this operator is used, the same description is used.

This can be better served using a static description for the operator, which would remove the need to copy the same description for every place the operator is used.

In fact, the description used in all instances of the operator is identical to the existing docstring, save for a missing newline character. So this patch adds that newline to the docstring and removes the dynamic description from use.