Fix: Sculpt Mode - Mesh - Options Panel - attribute not found error #5457
A bit of an educated guess, as doing this fixes the attribute issue.
Though whether `OPERATOR` is the correct value for `props.settings_source`, it a bit unknown.

The reason is was chosen is because it's what's used by `VIEW3D_MT_mask` from `space_view3d.py`, so it was assumed that `VIEW3D_PT_sculpt_options` from `space_view3d_toolbar.py` would similarly use that too.