3D View - Object - Parent sub menu - make entries consistent with the floating panel #5538 (#5601)
* 3D View - Object - Parent sub menu - make entries consistent with the floating panel #5538
This adds conditional systems and removes the access to the enum, making the menu lighter and more digestable based on parent selection context.

* Moved comment up and made other relevant areas with more explicit comments #5538

* Made disabled operators discoverable #5538
This also pushed the vertex group up to be more consistent with the context menu.

* Minimized code and made the disabled operators show in the header #5538
Allows to make this discoverable even if not relevant, while remaing grouped and useful.