3D View - Painting Mode - Make sidebar strength and radius control buttons consistent #5606 (#5610)
Here I added conditional visibility and a new dropdown panel to the modes that have relevant strength and size curves - so when you turn them on in the header, you see the drop down. Now it's consistent.

This is done by adding one extra drop down that is dynamic, and drawing the dropdown only when in header.

Note: many pressure curves are not applied to grease pencil like in the grease pencil vertex paint mode.

Now it's discoverable, and minimized, yet accessible.