Node Editors - Add group outlines for node groups #5627 (#5638)
This was done to replace the Blender solution for node groups: make them imperceptible by adding a footer indicator that can be confused with a drop shadow, dark on dark - and remove the icon. 

In turn this adds back the icon to imply it's still a group, with an additional highlight to quickly help the eye see what is the nodegroup about (header color) and quickly select and troubleshoot for general users. It also helps distinguish what is an asset and what is a built in node which should be clear for artists to know (instead of blindly trying to toggle built in nodes)

Note: If we are to adjust colors or outlines and padding in the future, let's do it in another task. 

# Commits:

* Initial outline for groups implemented as a new functions for node groups #5627
This replaced the belnder triple chin drop shadow goop.

* Added node header color to line #5627

* Added background higher up in the layers #5627
Currently decoupled from collapsed and uncollapsed, but has a slight selected/unselected transparency difference.

* Add back the icon for groups #5627
Also added the icon to the collapsed state consistently.

* Fix a compile warning of an unused element #5627