Update node_utils.py to fix TypeError when connecting Bool socket to Material Output
Issue: Connecting a Boolean-type socket to Material Output from within a NodeGroup using Ctrl+Shift+LeftClick caused a TypeError.

Cause: In node_utils.py, the socket type was incorrectly written as 'NodeSocketBoolean' instead of 'NodeSocketBool'.

Testing:
Verified the fix on Blender versions 4.2 through 5.0.1a. Only this instance of 'NodeSocketBoolean' existed in the codebase; all other references already used 'NodeSocketBool'.

Specifics: When connecting with Ctrl-Shift-LeftClick on a node inside of a nodegroup, a temporary socket is added to the the Group Output. When that socket was type BOOL, node_utils.py returned that the 'Viewer' socket created should be type NodeSocketBoolean, instead of NodeSocketBool.