Default Library - Make the library use a central library area to make future products better integrated #6009
Added a new utility script to copy files to a central library, and made a robust method to register and unregister the addon and clean things up, so this addon could be shipped and packed as seperate unique addons refereing to the same central libraries.

This will hopefully help finance the project and developers in the near future.