Default Library - Point central library to user folder instead of install folder for Appimage 5.0.1 release #6079
This moves the library to //USER/5.1/asset_library instead of the LOCAL core addon location, which can be read-only sometimes

This also makes new operators to add and remove libraries, and no-longer removes when addon deactivates. But it will add when addon activates. There is more library information too.

This also makes the registry of the addon somewhat more robust.
This did require a refactor... and though it has core principles for multiple addons to work in a modular way, this does not consider multiple registries of panels, operators and handlers. I will have to tackle that in another feature refactor task.