Merge fixes of the Default Library crash #6079 (#6100)
* Default Library - Minimize registration and warnings #6076

* Default Library - Minimize registration and warnings #6076
Made the fallback exist again..

Has a slight flicker, but makes it robust (this seems to be a blender 5.1 bug with library registrations not refreshing.) I will try minimize this for later.

* Default Library - Point central library to user folder instead of install folder for Appimage 5.0.1 release #6079

This moves the library to //USER/5.1/asset_library instead of the LOCAL core addon location, which can be read-only sometimes

This also makes new operators to add and remove libraries, and no-longer removes when addon deactivates. But it will add when addon activates. There is more library information too.

This also makes the registry of the addon somewhat more robust.
This did require a refactor... and though it has core principles for multiple addons to work in a modular way, this does not consider multiple registries of panels, operators and handlers. I will have to tackle that in another feature refactor task.

* Fixed a bad exception error #6079

* Minimized the registration and unregistration to be more stable on Mac and Linux #6079
This also minimized the refreshing.
Replaced the factory registry for a hopefully failsafe system for the submodules.

* Fix indentation error causing crash

* default library addon - fix remaining crashing when old existing default library were in userprefs

This should fix all issues, check on all 3 OS's and including Appimage on Linux and no errors messages appears or crashes were reproduced!.

---------

Co-authored-by: blenux <blenux@outlook.com>